Finished the Node Naming system. * We are going to get rid of (at least most of) the spec system. Simplified Logger Fixed boundbox calculation erorr in physics. Either we need to attach World25 and use by physicsmanager OR inherit. * I think use is  better. Added missing begin/endEdit in MeshUtils. first get stuff to render then fix frutum draw in cameranode Moved all drawing code to Scene (it was actually duplicated in PhysicsWorld). Implemented drawing routines in Script and also moved to Component (as component is used by SceneNode, not Script); Made World25 REFERENCE PhysicsManager. Changed debug key combos to use Shift (Shift+F1) Note now that we manually unload the physics world, we MUST remember to add back the scene objects to PhysicsWorld (somehow) or share their data with PhysicsWorld.

+++ src/bottle/W25MeshMaker.h
@@ -93,7 +93,7 @@ public:
 
   int numTrisSide(int side) {
     AssertOrThrow2(side >= 0 && side < c_nSides);
-    return _verts[side].size() / 3;
+    return _verts[side].size() / (size_t)3;
   }
   W25Geom getGeomId() {
     return _iGeomId;
